Split core openCV functions from numpy ones
